+++ src/buttons/mp_playlists.js
@@ -1,9 +1,11 @@
-const { handleList } = require("../utils/playlistPanelActions");
+const {
+  createPlaylistPanelMessage
+} = require("../utils/playlistPanel");
 
 module.exports = {
   customId: "mp_playlists",
 
   async execute(interaction) {
-    return handleList(interaction);
+    return interaction.update(createPlaylistPanelMessage());
   }
 };